helpers: split into enjs and dejs files
